MOD : Use genetic algorithm instead of simulated annealing : More stable, so probably more accurate
