+++ requirements.txt
@@ -6,6 +6,8 @@ sphinx==7.1.2
 sphinx-rtd-theme==1.3.0rc1
 packaging>=24.0
 GitPython>=3.0.0
+# == Genetic Algorithm ==
+scikit-opt>=0.6.0
 # == EMHASS ==
 numpy==1.26.4
 scipy==1.12.0
