+++ src/openhems/.vscode/settings.json
@@ -7,5 +7,6 @@
 		"test_*.py"
 	],
 	"python.testing.pytestEnabled": false,
-	"python.testing.unittestEnabled": true
+	"python.testing.unittestEnabled": true,
+	"python.pythonPath": "/home/alberic/Documents/OpenHomeSystem/openhems-sample/venvemhass/bin/python"
 }
